+++ packages/e2e-test-utils/CHANGELOG.md
@@ -2,6 +2,26 @@
 
 ## Unreleased
 
+## 11.19.0 (2025-02-28)
+
+## 11.18.0 (2025-02-12)
+
+## 11.17.0 (2025-01-29)
+
+## 11.16.0 (2025-01-15)
+
+## 11.15.0 (2025-01-02)
+
+## 11.14.0 (2024-12-11)
+
+## 11.13.0 (2024-11-27)
+
+## 11.12.0 (2024-11-16)
+
+## 11.11.0 (2024-10-30)
+
+## 11.10.0 (2024-10-16)
+
 ## 11.9.0 (2024-10-03)
 
 ## 11.8.0 (2024-09-19)
@@ -199,7 +219,7 @@
 
 ### Enhancements
 
--   `visitAdminPage` will now throw an error (emit a test failure) when there are unexpected errors on hte page.
+-   `visitAdminPage` will now throw an error (emit a test failure) when there are unexpected errors on the page.
 
 ### New Features